requirements.txt
appdirs==1.4.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
BeautifulSoup==3.2.1
beautifulsoup4==4.5.3
certifi==2017.1.23
decorator==4.0.11
Django==1.10.5
enum34==1.1.6
image==1.5.5
ipython==5.2.2
ipython-genutils==0.1.0
MySQL-python==1.2.5
olefile==0.44
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pyasn1==0.2.2
Pygments==2.2.0
pyparsing==2.1.10
rsa==3.4.2
scandir==1.4
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.5
tornado==4.5.2
traitlets==4.3.1
wcwidth==0.1.7
